Fixed case where there where holes in subplot spec
These cases, however, require ghost axes.  The alternative is to
keep track of axes within the layout manager.  Could do that if there
are serious consequences to having ghost axes.

Most of the time users probably use all the subplots in their gridspec,
or there is no ill effect to putting a ghost in.